+++ httpserver/src/main/kotlin/pt/isel/ps/g06/httpserver/dataAccess/input/ingredient/IngredientInput.kt
@@ -6,5 +6,5 @@ data class IngredientInput(
         @field:NotNull(message = "Ingredient identifier must not be empty!")
         val identifier: Int?,
         @field:NotNull(message = "An ingredient quantity must be given!")
-        val quantity: Int?
+        val quantity: Float?
 )
